Move a few more files into events src (#36574)
