+++ components/article/PlatformPicker.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 
 import { useArticleContext } from 'components/context/ArticleContext'
-import { parseUserAgent } from 'components/lib/user-agent'
+import { parseUserAgent } from 'src/events/components/user-agent'
 import { InArticlePicker } from './InArticlePicker'
 
 const platformQueryKey = 'platform'
