+++ lib/page.js
@@ -15,7 +15,7 @@ import { productMap } from './all-products.js'
 import slash from 'slash'
 import readFileContents from './read-file-contents.js'
 import getLinkData from './get-link-data.js'
-import getDocumentType from './get-document-type.js'
+import getDocumentType from '../src/events/lib/get-document-type.js'
 import { union } from 'lodash-es'
 import { allTools } from './all-tools.js'
 import { renderContentWithFallback } from './render-with-fallback.js'
