+++ src/events/components/Survey.module.scss
