+++ src/events/lib/get-document-type.js
