+++ src/events/lib/hydro.js
@@ -2,9 +2,9 @@ import { createHmac } from 'crypto'
 import { Agent } from 'node:https'
 import got from 'got'
 import { isNil } from 'lodash-es'
-import statsd from '../../lib/statsd.js'
-import { report } from '../../lib/failbot.js'
-import { MAX_REQUEST_TIMEOUT } from '../../lib/constants.js'
+import statsd from '../../../lib/statsd.js'
+import { report } from '../../../lib/failbot.js'
+import { MAX_REQUEST_TIMEOUT } from '../../../lib/constants.js'
 
 const TIME_OUT_TEXT = 'ms has passed since batch creation'
 const SERVER_DISCONNECT_TEXT = 'The server disconnected before a response was received'
