+++ src/events/lib/schema.js
@@ -1,7 +1,7 @@
-import { languageKeys } from '../../lib/languages.js'
-import { allVersionKeys } from '../../lib/all-versions.js'
-import { productIds } from '../../lib/all-products.js'
-import { allTools } from '../../lib/all-tools.js'
+import { languageKeys } from '../../../lib/languages.js'
+import { allVersionKeys } from '../../../lib/all-versions.js'
+import { productIds } from '../../../lib/all-products.js'
+import { allTools } from '../../../lib/all-tools.js'
 
 const versionPattern = '^\\d+(\\.\\d+)?(\\.\\d+)?$' // eslint-disable-line
