+++ src/search/components/SearchResults.tsx
@@ -7,7 +7,7 @@ import type { SearchResultsT, SearchResultHitT } from './types'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { Link } from 'components/Link'
 import { useQuery } from 'components/hooks/useQuery'
-import { sendEvent, EventType } from 'src/events/browser'
+import { sendEvent, EventType } from 'src/events/components/events'
 
 import styles from './SearchResults.module.scss'
